# x4 improvement on parsing + Fix the benchmarks