You can improve the parsing by 4 times by using a lookup table instead of an if/else branching to compute the hex value of a char.

```java
    private static int[] HEX_VALUE = new int[128];
    static {
        HEX_VALUE['0'] = 0;
        HEX_VALUE['1'] = 1;
        HEX_VALUE['2'] = 2;
        HEX_VALUE['3'] = 3;
        HEX_VALUE['4'] = 4;
        HEX_VALUE['5'] = 5;
        HEX_VALUE['6'] = 6;
        HEX_VALUE['7'] = 7;
        HEX_VALUE['8'] = 8;
        HEX_VALUE['9'] = 9;
        HEX_VALUE['a'] = 10;
        HEX_VALUE['A'] = 10;
        HEX_VALUE['b'] = 11;
        HEX_VALUE['B'] = 11;
        HEX_VALUE['c'] = 12;
        HEX_VALUE['C'] = 12;
        HEX_VALUE['d'] = 13;
        HEX_VALUE['D'] = 13;
        HEX_VALUE['e'] = 14;
        HEX_VALUE['E'] = 14;
        HEX_VALUE['f'] = 15;
        HEX_VALUE['F'] = 15;
    }
```

Then just use `HEX_VALUE[uuidSequence.charAt(X)];` instead of `getHexValueForChar(uuidSequence.charAt(X))`.

Benchmark pom project was fixed because it pointed to a `fast-uuid-parser` which is named `fast-uuid`.
I also removed some tests since the `getHexValueForChar` is now deleted.
Readme has been updated in order to explain how to run the benchmarks (for those who don't know jmh).

Benchmarks on my machine:

*Before*
```
Benchmark                                   Mode  Cnt         Score        Error  Units
UUIDBenchmark.benchmarkFastParserToString  thrpt   40  18850285,268 ± 292260,574  ops/s
UUIDBenchmark.benchmarkUUIDFromFastParser  thrpt   40   5430137,326 ± 354801,493  ops/s
UUIDBenchmark.benchmarkUUIDFromString      thrpt   40   1319052,901 ± 101011,389  ops/s
UUIDBenchmark.benchmarkUUIDToString        thrpt   40   2581791,612 ± 114785,630  ops/s
```
*After*
```
Benchmark                                   Mode  Cnt         Score        Error  Units
UUIDBenchmark.benchmarkFastParserToString  thrpt   40  18154539,206 ± 617354,677  ops/s
UUIDBenchmark.benchmarkUUIDFromFastParser  thrpt   40  23199948,837 ± 638756,033  ops/s
UUIDBenchmark.benchmarkUUIDFromString      thrpt   40   1359176,406 ±  44637,824  ops/s
UUIDBenchmark.benchmarkUUIDToString        thrpt   40   2468160,559 ± 158349,184  ops/s
```

Let's find some ways to make it even faster ;-)