@jchambers 2018-01-23
Awesome! You actually beat me to the punch here by just about ten minutes; I was literally writing exactly the same thing on my end. Our implementations are a little different, though, in that yours doesn't do any error-checking. That means it would, for example, parse `xvzjkpoi-mmmm-nssss-ghty-qwrtyuiopghj` as a valid (but entirely zeroed) UUID.

My approach was to populate the non-hex-digit slots in the array with `-1`, then check to see if the value we get from the array is non-negative (and throw an `IllegalArgumentException` otherwise). That does mean we're still doing one more check per character (and may mean we might want to keep the `getHexValueForChar` method), and so the gains aren't quite as pronounced. What do you think?

Regardless, it looks like I broke the benchmarks in 2a8b37bd2199bc10eee9f969b4d16a3e3d57d248; sorry for the trouble, and thanks for the fix!