@jchambers 2018-01-24
FWIW, I fixed the benchmarks in c1a79d6abece84fb4929c4a2870422b004c47fcf so that fix doesn't get held up by discussion here.