@jchambers 2018-01-24
@agrison I just pushed my approach to the same problem in #4. What do you think?