@jchambers 2018-01-24
I think the performance costs of doing some error-checking are small enough that it makes sense to go with the approach in #4. Thank you very, very much for this contribution!