@agrison 2018-01-24
Hey, I didn't notice you were from Boston, you're working overnight.. for me 😄 

You're right for the error handling, I actually had the same implementation, using `Arrays.fill(HEX_VALUE, -1)` and checking in `getHexValueForChar()` if it was `-1` in order to throw the exception like you did. On my machine I had something like a 15% boost by bypassing this check (avoiding a method call + condition) so I decided to push the code as is and let you decide what you really wanted. But I'm now reading that you only noticed a 5% boost so you're totally right to reactivate the error checking.

If it was for a 15% boost it would have been possible to let the user chose maybe using a boolean if he wants UUID validation also, but it seems a little YAGNI.

Regarding the `String`/`CharSequence` I was playing with `toCharArray` + direct access to the array instead of calling `charAt()` to avoid a method call and bounds checking that occurs in it, but it led to no performance improvement and then I forgot to revert that change, good catch 😄 

So now we're like 13 times faster on parsing than the UUID implementation, that's awesome !

Cheers 🏆