@jchambers 2018-01-24
> So now we're like 13 times faster on parsing than the UUID implementation, that's awesome !

14 times after 0409864 ;)

And, yes, I'm from Boston, but I'm actually traveling right now, so my schedule is a bit shifted (ha! bit-shifted) right now.